# docs/openapi.yaml
openapi: 3.1.0
info:
  title: XahauIndex API
  version: 1.0.0
  description: |
    Self-hostable metadata indexer for the Xahau network.
    Covers IOU tokens, URITokens, issuer profiles, DEX prices, and Hook activity.

    Token URLs use two path segments: `/v1/tokens/{currency}/{issuer}`.
    Pair URLs use `/v1/prices/{base}/{counter}` with issuer query params.
    Native asset code is `XAH`.
    No API key. Operators may enable a per-IP HTTP limit (`API_RATE_LIMIT_MAX`);
    over the limit the API returns `429` with `error.code` `RATE_LIMITED`.
  contact:
    url: https://github.com/PuppyTools/xahauindex-node
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local instance
...
components:
  parameters:
    page:
      name: page
      in: query
      description: Page number. Default 1
      schema:
        type: integer
        minimum: 1
        default: 1
    per_page:
      name: per_page
      in: query
      description: Rows per page (1–100). Default 20
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  schemas:
    Meta:
      type: object
      required: [ledger_index]
      properties:
        count:
          type: integer
        page:
          type: integer
        per_page:
          type: integer
        ledger_index:
          type: integer
          description: Ledger index at time of response
        history_start_ledger:
          type: integer
          description: Earliest ledger this node has price/transfer history for (backfill frontier while running, backfill_from when complete, else live_from_ledger)
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: NOT_FOUND
              description: "`NOT_FOUND`, `BAD_REQUEST`, `RATE_LIMITED`, or `INTERNAL`"
            message:
              type: string
...
    HookLabel:
      type: object
      required: [name, project, description, url, source]
      description: |
        Read-time catalog name. Hash match wins; if the hash is unknown the install
        account is checked (Evernode system accounts stay labeled after a hash rotation).
      properties:
        name:
          type: string
          example: Evernode heartbeat
        project:
          type: string
          example: Evernode
        description:
          type: string
        url:
          type: string
        source:
          type: string
          enum: [catalog]
    HookTriggerMask:
      type: object
      required: [raw, mode, types]
      properties:
        raw:
          type: string
        mode:
          type: string
          enum: [all_except, only, none]
        types:
          type: array
          items:
            type: string
    HookDefinition:
      type: object
      required: [hook_hash, code_size]
      properties:
        hook_hash:
          type: string
        label:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/HookLabel'
        hook_namespace:
          type: string
          nullable: true
        hook_on:
          type: string
          nullable: true
        hook_on_incoming:
          type: string
          nullable: true
        hook_on_outgoing:
          type: string
          nullable: true
        hook_can_emit:
          type: string
          nullable: true
        hook_name:
          type: string
          nullable: true
        hook_api_version:
          type: integer
          nullable: true
        parameters:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              value:
                type: string
        reference_count:
          type: integer
          nullable: true
        code_size:
          type: integer
        hook_fee:
          type: string
          nullable: true
        hook_callback_fee:
          type: string
          nullable: true
        hook_set_txn_id:
          type: string
          nullable: true
        flags:
          type: integer
          nullable: true
        triggers:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/HookTriggerMask'
        triggers_incoming:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/HookTriggerMask'
        triggers_outgoing:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/HookTriggerMask'
        can_emit:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/HookTriggerMask'
        first_ledger:
          type: integer
        last_updated:
          type: integer
    HookEntry:
      type: object
      required: [hook_hash]
      properties:
        hook_hash:
          type: string
          description: SHA-512Half of the WASM — the HookDefinition id
        label:
          nullable: true
          description: Static catalog name when the hash or Evernode system account is known
          allOf:
            - $ref: '#/components/schemas/HookLabel'
        hook_on:
          type: string
          nullable: true
        hook_on_incoming:
          type: string
          nullable: true
        hook_on_outgoing:
          type: string
          nullable: true
        hook_can_emit:
          type: string
          nullable: true
        hook_namespace:
          type: string
          nullable: true
        hook_name:
          type: string
          nullable: true
          description: NamedHooks trigger (4–16 bytes), not a product title
        hook_api_version:
          type: integer
          nullable: true
        parameters:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              value:
                type: string
        grants:
          type: array
          items:
            type: object
            properties:
              account:
                type: string
              hook_hash:
                type: string
        triggers:
          nullable: true
          description: Decoded HookOn (`all_except`, `only`, or `none`)
          allOf:
            - $ref: '#/components/schemas/HookTriggerMask'
        definition:
          nullable: true
          description: Shared defaults, `code_size`, fees. No bytecode
          allOf:
            - $ref: '#/components/schemas/HookDefinition'
    HookState:
      type: object
      required: [account, hook_count, hooks]
      properties:
        account:
          type: string
        hook_count:
          type: integer
        hooks:
          type: array
          items:
            $ref: '#/components/schemas/HookEntry'
        first_ledger:
          type: integer
        last_updated:
          type: integer
...
  examples:
    NotFound:
      value:
        error:
          code: NOT_FOUND
          message: Token USD/rHb9CJAWyB4rj91VRWn96DkukG4bwdtyTh not found
...
    HookAccounts:
      value:
        data:
          - account: rHktfGUbjqzU4GsYCMc1pDjdHXb5CJamto
            hook_count: 1
            hooks:
              - hook_hash: 1F7C84E14313C4FF2D4F39535428BF10767CCF8E87EFB51306CC3F94D13439EC
                label:
                  name: Evernode heartbeat
                  project: Evernode
                  description: Host heartbeats and reward distribution
                  url: https://docs.evernode.org/en/latest/platform/hooks/overview.html
                  source: catalog
                triggers:
                  raw: FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFBFFFFF
                  mode: none
                  types: []
                definition:
                  hook_hash: 1F7C84E14313C4FF2D4F39535428BF10767CCF8E87EFB51306CC3F94D13439EC
                  code_size: 500
                  reference_count: 12
                  hook_fee: "74"
                  label:
                    name: Evernode heartbeat
                    project: Evernode
                    description: Host heartbeats and reward distribution
                    url: https://docs.evernode.org/en/latest/platform/hooks/overview.html
                    source: catalog
            first_ledger: 100
            last_updated: 26272654
        meta:
          count: 1
          page: 1
          per_page: 5
          ledger_index: 26272654
    HookAccount:
      value:
        data:
          account: rHktfGUbjqzU4GsYCMc1pDjdHXb5CJamto
          hook_count: 1
          hooks:
            - hook_hash: 1F7C84E14313C4FF2D4F39535428BF10767CCF8E87EFB51306CC3F94D13439EC
              label:
                name: Evernode heartbeat
                project: Evernode
                description: Host heartbeats and reward distribution
                url: https://docs.evernode.org/en/latest/platform/hooks/overview.html
                source: catalog
              triggers:
                raw: FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFBFFFFF
                mode: none
                types: []
              definition:
                hook_hash: 1F7C84E14313C4FF2D4F39535428BF10767CCF8E87EFB51306CC3F94D13439EC
                code_size: 500
                reference_count: 12
                hook_fee: "74"
                label:
                  name: Evernode heartbeat
                  project: Evernode
                  description: Host heartbeats and reward distribution
                  url: https://docs.evernode.org/en/latest/platform/hooks/overview.html
                  source: catalog
          first_ledger: 100
          last_updated: 26272654
        meta:
          ledger_index: 26272654
    HookDefinition:
      value:
        data:
          hook_hash: 1F7C84E14313C4FF2D4F39535428BF10767CCF8E87EFB51306CC3F94D13439EC
          hook_namespace: null
          hook_on: FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFBFFFFF
          hook_on_incoming: null
          hook_on_outgoing: null
          hook_can_emit: null
          hook_name: null
          hook_api_version: 0
          parameters: []
          reference_count: 12
          code_size: 500
          hook_fee: "74"
          hook_callback_fee: null
          hook_set_txn_id: null
          flags: null
          triggers:
            raw: FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFBFFFFF
            mode: none
            types: []
          label:
            name: Evernode heartbeat
            project: Evernode
            description: Host heartbeats and reward distribution
            url: https://docs.evernode.org/en/latest/platform/hooks/overview.html
            source: catalog
          first_ledger: 100
          last_updated: 26272654
        meta:
          ledger_index: 26272654
    HookDefinitions:
      value:
        data:
          - hook_hash: 1F7C84E14313C4FF2D4F39535428BF10767CCF8E87EFB51306CC3F94D13439EC
            code_size: 500
            reference_count: 12
            hook_fee: "74"
            label:
              name: Evernode heartbeat
              project: Evernode
              description: Host heartbeats and reward distribution
              url: https://docs.evernode.org/en/latest/platform/hooks/overview.html
              source: catalog
        meta:
          page: 1
          per_page: 5
          ledger_index: 26272654
...
  responses:
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            not_found:
              $ref: '#/components/examples/NotFound'
...
paths:
...
  /v1/hooks/definitions:
    get:
      operationId: listHookDefinitions
      tags: [hooks]
      summary: List indexed HookDefinition objects
      description: Shared HookDefinition metadata (`code_size`, defaults, fees). Bytecode is not stored.
      x-try: /v1/hooks/definitions?per_page=5
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/per_page'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/HookDefinition'
                  meta:
                    $ref: '#/components/schemas/Meta'
              examples:
                definitions:
                  $ref: '#/components/examples/HookDefinitions'
  /v1/hooks/definitions/{hook_hash}:
    get:
      operationId: getHookDefinition
      tags: [hooks]
      summary: One HookDefinition by hash
      parameters:
        - name: hook_hash
          in: path
          required: true
          description: HookDefinition hash (64 hex characters)
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/HookDefinition'
                  meta:
                    $ref: '#/components/schemas/Meta'
              examples:
                definition:
                  $ref: '#/components/examples/HookDefinition'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/hooks/{account}:
    get:
      operationId: getHooks
      tags: [hooks]
      summary: Hook state for a Xahau account
      description: 404 if that account has no indexed Hook object.
      parameters:
        - name: account
          in: path
          required: true
          description: Account r-address
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/HookState'
                  meta:
                    $ref: '#/components/schemas/Meta'
              examples:
                hooks:
                  $ref: '#/components/examples/HookAccount'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/hooks:
    get:
      operationId: listHookAccounts
      tags: [hooks]
      summary: List accounts with an active Hook ledger object
      x-try: /v1/hooks?per_page=5
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/per_page'
        - name: hook_hash
          in: query
          description: Only accounts that currently install this hash
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/HookState'
                  meta:
                    $ref: '#/components/schemas/Meta'
              examples:
                accounts:
                  $ref: '#/components/examples/HookAccounts'
...